refactor: keep business logic in utils
