+++ tests/authentication.spec.ts
@@ -1,18 +1,23 @@
 import { test, expect } from './setup';
+import { login } from './utils';
 
 test.describe('Authentication', () => {
   test.beforeEach(async ({ page }) => {
     await page.goto('/');
   });
 
-  test('shows a Logout button after login', async ({ login, queries }) => {
-    await login();
+  test('shows a Logout button after login', async ({
+    page,
+    mockAgent,
+    queries,
+  }) => {
+    await login(page, mockAgent);
 
     expect(await queries.queryByText(/Logout/i)).toBeDefined();
   });
 
-  test('allows user to logout', async ({ login, queries }) => {
-    await login();
+  test('allows user to logout', async ({ page, mockAgent, queries }) => {
+    await login(page, mockAgent);
 
     const logoutButton = await queries.findByText(/Logout/i);
