+++ tests/utils.ts
@@ -1,8 +1,75 @@
 import type { Miniflare } from 'miniflare';
 import type { Page } from 'playwright-core';
 import type { MockAgent } from 'undici';
-import { queries } from '@playwright-testing-library/test';
+import { queries, getDocument } from '@playwright-testing-library/test';
 import type { Resource, ResourceMetadata } from '../worker/types';
+
+export async function login(
+  page: Page,
+  mockAgent: MockAgent,
+  name = 'edmundhung'
+) {
+  const $document = await getDocument(page);
+  const loginButton = await queries.findByText($document, /Login with GitHub/i);
+
+  const github = mockAgent.get('https://github.com');
+  const githubAPI = mockAgent.get('https://api.github.com');
+
+  github
+    .intercept({
+      path: '/login/oauth/access_token',
+      method: 'POST',
+    })
+    .reply(
+      200,
+      new URLSearchParams({
+        access_token: 'a-platform-for-sharing-everything-about-remix',
+        scope: 'emails',
+        token_type: 'bearer',
+      }).toString()
+    );
+
+  githubAPI
+    .intercept({
+      path: '/user',
+      method: 'GET',
+    })
+    .reply(200, {
+      id: 'dev',
+      login: name,
+      name: 'Remix Guide Developer',
+      email: 'dev@remix.guide',
+      avatar_url: null,
+    });
+
+  await page.route('/login', async (route, request) => {
+    // It seems like there is no way to intercept request in the middle of the redirect
+    // Due to limitation of the devtool protocol
+    // We expect the browser is already redirected to github login page after the request
+    const response = await page.request.fetch(request);
+    const responseURL = new URL(response.url());
+    const returnTo = responseURL.searchParams.get('return_to');
+
+    // We need the `return_to` search params to find out all information we need
+    if (!returnTo) {
+      await route.abort();
+      return;
+    }
+
+    const url = new URL(decodeURIComponent(returnTo), responseURL.origin);
+
+    route.fulfill({
+      status: 302,
+      headers: {
+        Location: `${url.searchParams.get(
+          'redirect_uri'
+        )}?code=remix-guide&state=${url.searchParams.get('state')}`,
+      },
+    });
+  });
+
+  await loginButton.click();
+}
 
 export async function submitURL(page: Page, url: string, category = 'others') {
   const $form = await page.$('form[action="/submit"]');
